Add files via upload
This project was done using Python. Specifically, I used machine learning based classification algorithms to study the credit card crisis of Taiwan. The primary purpose of this paper was to determine which variables of the given dataset were the strongest predictors of default payment. The tests and models that were used to determine these variables were Random Forest Classification, Naive Bayes Classification and Gradient Boosting. The corresponding ".pdf" file is enclosed.